Bot Telegram/requirements.txt
apscheduler==3.6.3
backports.zoneinfo==0.2.1; python_version < '3.9'
cachetools==4.2.2
certifi==2022.6.15
python-telegram-bot==13.13
pytz-deprecation-shim==0.1.0.post0
pytz==2022.1
six==1.16.0
tornado==6.2
tzdata==2022.1; python_version >= '3.6'
tzlocal==4.2
